[Feat] vegi add calender로 분리, 완료 버튼 색상 관리 구현 (#129)
